pyproject.toml
[tool.poetry]
name = "Introduction_Airflow"
version = "0.1.0"
description = "An Introduction project to Airflow"
authors = ["Preethi Prakash <preethi93@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.8"
apache-airflow = "2.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
